Extract stream extensions to their own file.
