+++ OpenRA.FileFormats/OpenRA.FileFormats.csproj
@@ -141,6 +141,7 @@
     <Compile Include="Graphics\ShpTSReader.cs" />
     <Compile Include="FileFormats\XccLocalDatabase.cs" />
     <Compile Include="FileFormats\XccGlobalDatabase.cs" />
+    <Compile Include="StreamExts.cs" />
   </ItemGroup>
   <ItemGroup>
     <BootstrapperPackage Include="Microsoft.Net.Client.3.5">
